changed to avoid warnings from intel c++ compiler
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@4395 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ SelectedOutput.cpp
@@ -62,12 +62,12 @@ CVar CSelectedOutput::Get(int nRow, int nCol)const
 
 VRESULT CSelectedOutput::Get(int nRow, int nCol, VAR* pVAR)const
 {
-	if ((size_t)nRow >= this->GetRowCount() || (size_t)nRow < 0) {
+	if ((size_t)nRow >= this->GetRowCount() || nRow < 0) {
 		pVAR->type = TT_ERROR;
 		pVAR->vresult = VR_INVALIDROW;
 		return pVAR->vresult;
 	}
-	if ((size_t)nCol >= this->GetColCount() || (size_t)nCol < 0) {
+	if ((size_t)nCol >= this->GetColCount() || nCol < 0) {
 		pVAR->type = TT_ERROR;
 		pVAR->vresult = VR_INVALIDCOL;
 		return pVAR->vresult;
